+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -101,6 +101,7 @@ class RichContentEditor extends Component {
       config,
       isMobile,
       setEditorState: this.setEditorState,
+      getEditorState: this.getEditorState,
       getEditorBounds: this.getEditorBounds,
       languageDir: getLangDir(locale),
       shouldRenderOptimizedImages,
@@ -108,7 +109,6 @@ class RichContentEditor extends Component {
       siteDomain,
       setInPluginEditingMode: this.setInPluginEditingMode,
       getInPluginEditingMode: this.getInPluginEditingMode,
-      getEditorState: this.getEditorState,
     };
   };