+++ packages/plugin-image/web/src/image-component.jsx
@@ -189,9 +189,9 @@ ImageComponent.propTypes = {
   helpers: PropTypes.object.isRequired,
   t: PropTypes.func.isRequired,
   theme: PropTypes.object.isRequired,
-  disableRightClick: PropTypes.bool.isRequired,
-  getInPluginEditingMode: PropTypes.func.isRequired,
-  setInPluginEditingMode: PropTypes.func.isRequired,
+  disableRightClick: PropTypes.bool,
+  getInPluginEditingMode: PropTypes.func,
+  setInPluginEditingMode: PropTypes.func,
   isMobile: PropTypes.bool.isRequired,
 };