+++ CHANGELOG.md
@@ -12,7 +12,13 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-  
+
+  ### :bug: Bug Fix
+  - `image`
+    - [#657](https://github.com/wix-incubator/rich-content/pull/657) defaults for editor-only props in image-viewer
+    - [#657](https://github.com/wix-incubator/rich-content/pull/657) in-plugin editing input rendered in edit mode only
+    - [#657](https://github.com/wix-incubator/rich-content/pull/657) highres URL on SSR protection
+
 </details>
 <hr/>