+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   </summary>
 
   ### :bug: Bug Fix
+  - `common`
+    - [#658](https://github.com/wix-incubator/rich-content/pull/658) `types` definitions moved to `src/consts` from `statics`
   - `image`
     - [#657](https://github.com/wix-incubator/rich-content/pull/657) defaults for editor-only props in image-viewer
     - [#657](https://github.com/wix-incubator/rich-content/pull/657) in-plugin editing input rendered in edit mode only