+++ CHANGELOG.md
@@ -13,9 +13,24 @@
     Click to see more.
   </summary>
 
-
 </details>
 <hr/>
+
+## 6.8.2 (Feb 9, 2020)
+### :bug: Bug Fix
+- `common`
+  - [#658](https://github.com/wix-incubator/rich-content/pull/658) `types` definitions moved to `src/consts` from `statics`
+- `image`
+  - [#657](https://github.com/wix-incubator/rich-content/pull/657) defaults for editor-only props in image-viewer
+  - [#657](https://github.com/wix-incubator/rich-content/pull/657) in-plugin editing input rendered in edit mode only
+  - [#657](https://github.com/wix-incubator/rich-content/pull/657) highres URL on SSR protection
+
+
+
+## 6.8.1 (Feb 6, 2020)
+### :bug: Bug Fix
+- `viewer`
+  - [#654](https://github.com/wix-incubator/rich-content/pull/654) plugin settings retrieved correctly
 
 ## 6.8.0 (Feb 5, 2020)
 ### :rocket: New Feature